Fix auto-dosing stall + MET sensor fixes from code review
- AUTO_STABILISING: stall timeout (10min) now feeds the existing
  ineffective-dose/alarm accounting instead of bypassing it, so a probe
  that's flaky specifically during stabilising still reaches
  ALARM_REASON_NO_EC_RESPONSE instead of dosing forever unnoticed
- AUTO_SAMPLING: new stall guard (20min) + ALARM_REASON_EC_DATA_UNAVAILABLE
  surfaces a probe that never produces usable data instead of idling
  silently forever with no visibility
- MET: sensors.amb presence flag now includes MET, ambient fallback now
  checks this tick's actual Ambient read result (ambOk) instead of just
  sensor presence, boot detection retries 3x, lux register merged into
  the main block read (one Modbus transaction instead of two)
- cloud.cpp: MET backfill upload formatting now matches the live MQTT
  path's (int) casts for wind direction/PM2.5/PM10
- loopTask stack/heap log: heartbeat widened 60s->5min, kept instant
  logging on any new low-water mark so a real crash-class regression
  isn't lost to the reboot before the next heartbeat

Version intentionally left at 1.2.9 - that version was never tagged or
released, so these are fixes to it, not a new release.

Not yet bench-tested. Needs office-unit validation (deliberately force
an EC probe failure mid-STABILISING and at boot) before any tag/OTA.

+++ src/cloud.cpp
@@ -350,7 +350,7 @@ void uploadSensorReadings()
     JsonObject wd = arr.createNestedObject();
     wd["device"]    = deviceName;
     wd["sensor_id"] = wdId;
-    wd["value"]     = sensors.windDir;
+    wd["value"]     = (int)sensors.windDir; // matches the live MQTT `met.wd` (int) cast
 
     JsonObject no = arr.createNestedObject();
     no["device"]    = deviceName;
@@ -360,12 +360,12 @@ void uploadSensorReadings()
     JsonObject p1 = arr.createNestedObject();
     p1["device"]    = deviceName;
     p1["sensor_id"] = p1Id;
-    p1["value"]     = sensors.pm25;
+    p1["value"]     = (int)sensors.pm25; // matches the live MQTT `met.pm25` (int) cast
 
     JsonObject p2 = arr.createNestedObject();
     p2["device"]    = deviceName;
     p2["sensor_id"] = p2Id;
-    p2["value"]     = sensors.pm10;
+    p2["value"]     = (int)sensors.pm10; // matches the live MQTT `met.pm10` (int) cast
   }
 
   String rec = isoNow();
